content-visibility: Ensure to keep the DL unlocked if targeted by scroll
This patch generalizes the find-in-page fix we had previously to also
apply to other methods that can use to scroll the viewport to the
element in a locked subtree. Specifically, this keep the lock unlocked
for two frames to ensure that the visibility notification kicks in to
keep it unlocked until the user scrolls away.

R=chrishtr@chromium.org

Bug: 1132869
Change-Id: I598f82f73429e41839a6675321eb6fb1eb84f6b5